+++ CHANGELOG.md
@@ -12,6 +12,9 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 ### Changed
 
+- Fix issue when updating roles with empty values (cf. [#147](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/147))
+- Support of entitlements with attribute for dimension import
+
 ### Fixed
 
 ### Removed